platformize Item & List
