+++ src/components/List/Item/ArchiveAction.tsx
@@ -3,20 +3,21 @@ import { Archive } from 'react-feather';
 import { useQueryClient } from 'react-query';
 
 import { Action } from 'components/Action';
-import { archiveItem, queryKeys } from 'services/pocket/api';
+import { useListContext } from 'components/List';
 import { clearCache } from 'utils/dataCache';
 import { useItemContext } from './ItemContext';
 
 export const ArchiveAction: FC = () => {
   const queryClient = useQueryClient();
-  const { rgb, itemId } = useItemContext();
+  const { getQueryKey, archiveItem } = useListContext();
+  const { rgb, id } = useItemContext();
   const [archiveLoading, setArchiveLoading] = useState(false);
 
   const onArchive = async () => {
     setArchiveLoading(true);
-    const ok = await archiveItem(itemId);
+    const ok = await archiveItem(id);
     if (ok) {
-      await clearCache(queryKeys.get, queryClient);
+      await clearCache(getQueryKey, queryClient);
     }
     setArchiveLoading(false);
   };
