+++ src/components/List/Item/DeleteAction.tsx
@@ -3,20 +3,21 @@ import { Trash2 } from 'react-feather';
 import { useQueryClient } from 'react-query';
 
 import { Action } from 'components/Action';
-import { deleteItem, queryKeys } from 'services/pocket/api';
+import { useListContext } from 'components/List';
 import { clearCache } from 'utils/dataCache';
 import { useItemContext } from './ItemContext';
 
 export const DeleteAction: FC = () => {
   const queryClient = useQueryClient();
-  const { rgb, itemId } = useItemContext();
+  const { getQueryKey, deleteItem } = useListContext();
+  const { rgb, id } = useItemContext();
   const [deleteLoading, setDeleteLoading] = useState(false);
 
   const onDelete = async () => {
     setDeleteLoading(true);
-    const ok = await deleteItem(itemId);
+    const ok = await deleteItem(id);
     if (ok) {
-      await clearCache(queryKeys.get, queryClient);
+      await clearCache(getQueryKey, queryClient);
     }
     setDeleteLoading(false);
   };
